feat(example): bump to rn 0.74 & add react-navigation
