+++ example/package.json
@@ -9,11 +9,15 @@
     "pods": "pod-install --quiet"
   },
   "dependencies": {
-    "@react-native/babel-preset": "^0.73.20",
+    "@react-native/babel-preset": "0.74.88",
     "@react-native/metro-config": "^0.74.79",
+    "@react-navigation/native": "^6.1.18",
+    "@react-navigation/native-stack": "^6.11.0",
     "react": "18.2.0",
-    "react-native": "0.73.2",
+    "react-native": "0.74.6",
     "react-native-fs": "^2.20.0",
+    "react-native-safe-area-context": "^4.14.0",
+    "react-native-screens": "~3.32.0",
     "react-native-sound": "^0.11.2",
     "react-native-zip-archive": "^7.0.1"
   },
